BREAKING CHANGE: moving IteratorStreams from package utils to package streams
