+++ src/main/java/com/coreoz/windmill/files/FileType.java
@@ -1,7 +1,7 @@
 package com.coreoz.windmill.files;
 
 public enum FileType {
-	
+
 	ZIP, CFBF, UNKNOWN
 
 }
