+++ src/main/java/com/coreoz/windmill/imports/parsers/csv/CsvRowIterator.java
@@ -5,7 +5,7 @@ import java.util.stream.Collectors;
 
 import com.coreoz.windmill.imports.FileSchema;
 import com.coreoz.windmill.imports.Row;
-import com.coreoz.windmill.utils.IteratorStreams;
+import com.coreoz.windmill.streams.IteratorStreams;
 
 class CsvRowIterator implements Iterator<Row> {
